refactor(store): extract _IntegrityMixin to _integrity.py (#280)
Fourth step of the VstashStore decomposition (#280). Move-only: the 2
integrity methods (integrity_check, integrity_repair) move verbatim into
_IntegrityMixin in vstash/store/_integrity.py; VstashStore inherits it. The
now-unused IntegrityCheck/IntegrityRepair model imports are dropped from
__init__.py (they were only used by these methods; callers import them from
vstash.models, not vstash.store).

Method bodies are byte-identical (verified via diff). Cross-mixin calls
(_bump_cache_epoch, _invalidate_idf_cache) keep resolving via MRO.

- pytest tests/ -> 1253 passed (1251 + 2 new test_integrity_manager.py), 6 deselected
- ruff check/format clean; MRO adds _IntegrityMixin after _IndexBackendMixin
